+++ src/onegov.town/onegov/town/assets/js/editor.js
@@ -1,9 +1,12 @@
 $(function() {
     _.each($('textarea.editor'), function(el) {
         var textarea = $(el);
+        var form = textarea.closest('form');
 
         textarea.redactor({
-            imageUpload: textarea.closest('form').data('image-upload-url')
+            imageUpload: form.data('image-upload-url'),
+            imageManagerJson: form.data('image-list-url'),
+            plugins: ['imagemanager']
         });
     });
 });