+++ src/onegov.town/onegov/town/tests/test_views.py
@@ -4,6 +4,7 @@ import onegov.town
 
 from mock import patch
 from morepath import setup
+from onegov.testing import utils
 from webtest import TestApp as Client
 from webtest import Upload
 
@@ -90,13 +91,27 @@ def test_view_images(town_app):
 
     assert "Noch keine Bilder hochgeladen" in images_page
 
-    images_page.form['file'] = Upload('Test.txt', b'x')
+    images_page.form['file'] = Upload('Test.txt', utils.create_image().read())
     assert images_page.form.submit(expect_errors=True).status_code == 415
 
-    images_page.form['file'] = Upload('Test.jpg', b'x')
+    images_page.form['file'] = Upload('Test.jpg', utils.create_image().read())
     images_page = images_page.form.submit().follow()
 
     assert "Noch keine Bilder hochgeladen" not in images_page
+
+    name = images_page.pyquery('img').attr('src').split('/')[-1]
+
+    # thumbnails are created on the fly
+    assert town_app.filestorage.exists('images/' + name)
+    assert not town_app.filestorage.exists('images/thumbnails' + name)
+    client.get('/thumbnails/' + name)
+    assert town_app.filestorage.exists('images/thumbnails/' + name)
+
+    # thumbnails are deleted with the image
+    delete_link = images_page.pyquery('a.delete').attr('ic-delete-from')
+    client.delete(delete_link)
+    assert not town_app.filestorage.exists('images/' + name)
+    assert not town_app.filestorage.exists('images/thumbnails/' + name)
 
 
 def test_startpage(town_app):