+++ src/onegov.town/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+0.0.2 (2015-05-05)
+~~~~~~~~~~~~~~~~~~~
+
 - Images are now always shown in order of their creation.
   [href]